+++ class/Type/Decimal.php
@@ -2,7 +2,7 @@
 /**
  * This file is part of a "Docalist Core" plugin.
  *
- * Copyright (C) 2012-2014 Daniel Ménard
+ * Copyright (C) 2012-2015 Daniel Ménard
  *
  * For copyright and license information, please view the
  * LICENSE.txt file that was distributed with this source code.
@@ -18,7 +18,7 @@ use Docalist\Type\Exception\InvalidTypeException;
 /**
  * Type nombre décimal.
  */
-class Float extends Scalar {
+class Decimal extends Scalar {
     static protected $default = 0.0;
 
     public function assign($value) {
